WIP-v2-Extracted feature into ManagementController and added first passing spec test
