--- app/views/layouts/modal.html.erb
@@ -1,14 +1,0 @@
-<div class="modal" id="mainModal" tabindex="-1" role="dialog" aria-labelledby="mainModalLabel" aria-hidden="true">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span><span class="sr-only">Close</span></button>
-        <h4 class="modal-title" id="mainModalLabel">
-          <%= yield :title if content_for? :title %>&nbsp;
-        </h4>
-      </div>
-
-      <%= yield %>
-    </div>
-  </div>
-</div>
