+++ app/views/management/edit_student_questions.html.erb
@@ -1,4 +1,4 @@
-<%= form_tag delete_answered_questions_path, url: { action: 'delete_answered_questions' }, method: :delete do |f| %>
+<%= form_tag delete_student_questions_path, url: { action: 'delete_student_questions' }, method: :delete do |f| %>
 
   <% @answered_questions.each do |answered_question| %>
     <div class="lesson-question-wrapper">
