+++ app/views/management/student_manager.html.erb
@@ -1,0 +1,89 @@
+<div>
+  <%= form_tag (get_student_management_path) do %>
+    <%= label_tag 'email', 'Email' %>
+    <%= text_field_tag 'email' %>
+    <%= submit_tag "Get Student Questions"%>
+  <% end %>
+</div>
+
+<% if !!@records %>
+  <%= @records[:courses].each do |course| %>
+    <div class="container" >
+      <h4>Course: <%= course.name %></h4>
+      <% @records[:units].each do |unit| %>
+        <% if unit.course_id == course.id %>
+          <div class="answered-questions">
+            <h5>Unit: <%= unit.name %></h4>
+            <% @records[:topics].each do |topic| %>
+              <% if topic.unit_id == unit.id %>
+                <div class="lesson-question-wrapper">
+                  <ul>
+                    <li>
+                      Chapter: <%= topic.name %> |
+                      Exp: <%= get_topic_exp(topic.id, @records[:student_id]) %> |
+                      <!-- Trigger/Open The Modal -->
+                      <button id="topic-<%= topic.id %>" class="expBtn">Edit Exp</button>
+
+                      <!-- The Modal -->
+                      <div id="expModal-topic-<%= topic.id %>" class="modal">
+
+                        <!-- Modal content -->
+                        <div class="modal-content">
+                          <span class="closeModal close">x</span>
+                          <p>
+                            <%= form_var = form_tag (edit_experience_path) do %>
+                              <%= label_tag :exp, 'Experience' %>
+                              <%= text_field_tag :exp, get_topic_exp(topic.id, @records[:student_id]) %>
+                              <%= hidden_field_tag :id, topic.id %>
+                              <%= hidden_field_tag :student_id, @records[:student_id] %>
+                              <%= hidden_field_tag :exp_type, 'topic' %>
+                              <%= submit_tag 'Update experience' %>
+                            <% end %>
+                          </p>
+                        </div>
+
+                      </div>
+                      <!-- MODAL END -->
+                      <ul>
+                        <% @records[:lessons].each do |lesson| %>
+                          <% if lesson.topic_id == topic.id %>
+                            <li>
+                              Lesson: <%= lesson.name %> | Exp: <%= get_lesson_exp(lesson.id, @records[:student_id]) %>  / <%= lesson.pass_experience%> |
+                              <!-- Trigger/Open The Modal -->
+                              <button id="lesson-<%= lesson.id %>" class="expBtn">Edit Exp</button>
+
+                              <!-- The Modal -->
+                              <div id="expModal-lesson-<%= lesson.id %>" class="modal">
+
+                                <!-- Modal content -->
+                                <div class="modal-content">
+                                  <span class="closeModal close">x</span>
+                                  <p>
+                                    <%= form_var = form_tag (edit_experience_path) do %>
+                                      <%= label_tag :exp, 'Experience' %>
+                                      <%= text_field_tag :exp, get_lesson_exp(lesson.id, @records[:student_id]) %>
+                                      <%= hidden_field_tag :id, lesson.id %>
+                                      <%= hidden_field_tag :student_id, @records[:student_id] %>
+                                      <%= hidden_field_tag :exp_type, 'lesson' %>
+                                      <%= submit_tag 'Update experience' %>
+                                    <% end %>
+                                  </p>
+                                </div>
+                              </div>
+                              <!-- MODAL END -->
+                              | <%= link_to "Edit Answered Questions", edit_student_questions_path(lesson_id: lesson.id, user_id: @records[:student_id]), id:"edit-answered-question-lesson-#{lesson.id}" %>
+                            </li>
+                          <% end %>
+                        <% end %>
+                      </ul>
+                    </li>
+                  </ul>
+                </div>
+              <% end %>
+            <% end %>
+          </div>
+        </div>
+        <% end %>
+      <% end %>
+  <% end %>
+<% end %>
