+++ app/views/shared/_header.html.erb
@@ -8,6 +8,7 @@
       <li><%= link_to 'Questions', questions_path %></li>
       <li><%= link_to 'Add Question', new_question_path %></li>
       <li><%= link_to 'Student Questions', answered_questions_path %></li>
+      <li><%= link_to 'Student Manager', student_manager_path %></li>
     <% end %>
     <li><%= link_to 'About', about_path %></li>
     <li><%= link_to 'FAQ', faq_path %></li>
